feat: 기존 테스트 페이지 삭제 및 쿼리 받아오기 확인 (#33)
