+++ src/components/layouts/search-bar.tsx
@@ -1,46 +1,33 @@
+'use client';
+
 import { Command, CommandInput } from '@/components/ui/command';
-// import {
-//   CommandEmpty,
-//   CommandGroup,
-//   CommandItem,
-//   CommandList,
-//   CommandSeparator,
-//   CommandShortcut,
-// } from '@/components/ui/command';
+import { useRouter } from 'next/navigation';
+import { useRef } from 'react';
 
 const SearchBar = () => {
+  const inputRef = useRef<HTMLInputElement>(null);
+  const router = useRouter();
+
+  const handleSearch = () => {
+    const query = inputRef.current?.value;
+    if (query) {
+      router.push(`/search?query=${query}`);
+    }
+  };
+
   return (
     <>
       <Command className="w-[300px] rounded-lg border">
-        <CommandInput placeholder="검색어를 입력해 주세요" />
-        {/* <CommandList>
-          <CommandEmpty>No results found.</CommandEmpty>
-          <CommandGroup heading="Suggestions">
-            <CommandItem>
-              <span>Calendar</span>
-            </CommandItem>
-            <CommandItem>
-              <span>Search Emoji</span>
-            </CommandItem>
-            <CommandItem disabled>
-              <span>Calculator</span>
-            </CommandItem>
-          </CommandGroup>
-          <CommandSeparator />
-          <CommandGroup heading="Settings">
-            <CommandItem>
-              <span>Profile</span>
-            </CommandItem>
-            <CommandItem>
-              <span>Billing</span>
-              <CommandShortcut>⌘B</CommandShortcut>
-            </CommandItem>
-            <CommandItem>
-              <span>Settings</span>
-              <CommandShortcut>⌘S</CommandShortcut>
-            </CommandItem>
-          </CommandGroup>
-        </CommandList> */}
+        <CommandInput
+          placeholder="검색어를 입력해 주세요"
+          ref={inputRef}
+          onKeyDown={(e) => {
+            if (e.key === 'Enter') {
+              e.preventDefault();
+              handleSearch();
+            }
+          }}
+        />
       </Command>
     </>
   );
